Add react-router switch to index.js Move appbar to index.js Add yet to be implemented route structure for "rooms"
